Add empty on_result function parameter in rtt-poll request
In dap.nvim, when on_result is nil, behaviour is to add a coroutine and on_result that resume it.
In this case, it yields just after call, that was causing 'yield accross C boundary error'. Fixed
with an empty on_result function on call session:request('rtt-poll').

Tested with a Zephyr project using RTT console, working properly